refactor model helpers
