--- mri_degad/workflow/scripts/model_helpers/src/.gitignore
@@ -1,221 +1,0 @@
-# Python
-__pycache__/
-*.py[cod]
-*$py.class
-*.so
-.Python
-build/
-develop-eggs/
-dist/
-downloads/
-eggs/
-.eggs/
-lib/
-lib64/
-parts/
-sdist/
-var/
-wheels/
-share/python-wheels/
-*.egg-info/
-.installed.cfg
-*.egg
-MANIFEST
-
-# PyInstaller
-*.manifest
-*.spec
-
-# Installer logs
-pip-log.txt
-pip-delete-this-directory.txt
-
-# Unit test / coverage reports
-htmlcov/
-.tox/
-.nox/
-.coverage
-.coverage.*
-.cache
-nosetests.xml
-coverage.xml
-*.cover
-*.py,cover
-.hypothesis/
-.pytest_cache/
-cover/
-
-# Translations
-*.mo
-*.pot
-
-# Django stuff:
-*.log
-local_settings.py
-db.sqlite3
-db.sqlite3-journal
-
-# Flask stuff:
-instance/
-.webassets-cache
-
-# Scrapy stuff:
-.scrapy
-
-# Sphinx documentation
-docs/_build/
-
-# PyBuilder
-.pybuilder/
-target/
-
-# Jupyter Notebook
-.ipynb_checkpoints
-
-# IPython
-profile_default/
-ipython_config.py
-
-# pyenv
-.python-version
-
-# pipenv
-Pipfile.lock
-
-# poetry
-poetry.lock
-
-# pdm
-.pdm.toml
-
-# PEP 582
-__pypackages__/
-
-# Celery stuff
-celerybeat-schedule
-celerybeat.pid
-
-# SageMath parsed files
-*.sage.py
-
-# Environments
-.env
-.venv
-env/
-venv/
-ENV/
-env.bak/
-venv.bak/
-
-# Spyder project settings
-.spyderproject
-.spyproject
-
-# Rope project settings
-.ropeproject
-
-# mkdocs documentation
-/site
-
-# mypy
-.mypy_cache/
-.dmypy.json
-dmypy.json
-
-# Pyre type checker
-.pyre/
-
-# pytype static type analyzer
-.pytype/
-
-# Cython debug symbols
-cython_debug/
-
-# PyCharm
-.idea/
-
-# VS Code
-.vscode/
-
-# Temporary files
-*.tmp
-*.temp
-*.swp
-*.swo
-*~
-
-# OS generated files
-.DS_Store
-.DS_Store?
-._*
-.Spotlight-V100
-.Trashes
-ehthumbs.db
-Thumbs.db
-
-# NIfTI files
-*.nii
-*.nii.gz
-*.hdr
-*.img
-*.hdr.gz
-*.img.gz
-
-# Medical imaging formats
-*.dcm
-*.dicom
-*.mha
-*.mhd
-*.raw
-
-# Data files
-*.csv
-*.json
-*.pkl
-*.pickle
-*.h5
-*.hdf5
-*.mat
-
-# Model files
-*.pth
-*.pt
-*.ckpt
-*.safetensors
-
-# Logs
-*.log
-logs/
-tensorboard_logs/
-
-# Checkpoints and experiments
-checkpoints/
-experiments/
-runs/
-wandb/
-
-# Cache directories
-.cache/
-cache/
-__pycache__/
-
-# Temporary directories
-tmp/
-temp/
-.tmp/
-
-# Large data files
-*.zip
-*.tar.gz
-*.rar
-*.7z
-
-# Backup files
-*.bak
-*.backup
-*.old
-
-# Configuration files with sensitive data
-config_local.json
-secrets.json
-.env.local
