--- mri_degad/workflow/scripts/model_helpers/src/utils/adaptive_losses.py
@@ -1,439 +1,0 @@
-import numpy as np
-import warnings
-import torch
-from tqdm import tqdm
-import matplotlib.pyplot as plt
-import os
-from typing import List, Tuple, Optional, Union, Callable
-import logging
-
-# Set up logging
-logging.basicConfig(level=logging.INFO)
-logger = logging.getLogger(__name__)
-
-class AdaMultiLossesNormTorch:
-    def __init__(self, eps: float = 1e-8):
-        """
-        Initialize the adaptive multi-loss normalization class.
-        
-        Args:
-            eps (float): Small epsilon value to prevent division by zero
-        """
-        self.eps = eps
-    
-    def compute_losses(self, dataset, batch_size: Optional[int] = None, 
-                      losses: Optional[List[Callable]] = None, device: str = 'cpu') -> torch.Tensor:
-        """
-        Compute multiple losses over a dataset with optimized batch processing.
-        Returns PyTorch tensor instead of NumPy array for better performance.
-
-        Args:
-            dataset (torch.utils.data.DataLoader): The PyTorch DataLoader to iterate over.
-            batch_size (int, optional): The expected batch size. If provided, incomplete
-                                        batches at the end of the dataset are skipped.
-            losses (list of callable): PyTorch loss functions that take (output, target) as input.
-            device (str): Device to move tensors to ('cpu' or 'cuda').
-
-        Returns:
-            torch.Tensor: A PyTorch tensor where each row corresponds to the collected losses
-                         for one of the loss functions.
-        """
-        if losses is None:
-            losses = []
-
-        if not losses:
-            return torch.empty(0, 0, device=device)
-
-        # Pre-allocate memory for better performance
-        collected_losses = []
-        num_losses = len(losses)
-        
-        # Use tqdm for progress tracking
-        pbar = tqdm(dataset, desc="Computing initial losses")
-        
-        for batch_data in pbar:
-            motion_data, free = batch_data
-            
-            # Handle cases where motion_data is a tuple of tensors (e.g., for SAP)
-            if isinstance(motion_data, (list, tuple)):
-                # Assuming the main motion slice is the middle one if SAP is enabled
-                motion = motion_data[1] 
-            else:
-                motion = motion_data
-
-            # Optimize tensor operations - only add batch dimension if needed
-            if motion.dim() == 3:  # (C, H, W)
-                motion = motion.unsqueeze(0)  # (1, C, H, W)
-            if free.dim() == 3:  # (C, H, W)
-                free = free.unsqueeze(0)  # (1, C, H, W)
-            
-            # Check batch size if specified - do this before device transfer
-            if batch_size and (free.shape[0] != batch_size or motion.shape[0] != batch_size):
-                logger.debug(f"Skipping incomplete batch. Free: {free.shape}, Motion: {motion.shape}")
-                continue
-            
-            # Move tensors to device only once
-            motion = motion.to(device, non_blocking=True)
-            free = free.to(device, non_blocking=True)
-            
-            # Compute all losses in one pass with torch.no_grad()
-            with torch.no_grad():
-                batch_losses = torch.zeros(num_losses, device=device)
-                for i, loss_fn in enumerate(losses):
-                    batch_losses[i] = loss_fn(motion, free)
-            
-            collected_losses.append(batch_losses)
-        
-        if not collected_losses:
-            return torch.empty(len(losses), 0, device=device)
-
-        # Stack all batch results efficiently - keep as PyTorch tensor
-        return torch.stack(collected_losses, dim=1)  # Shape: (num_losses, num_batches)
-
-    def compute_normalized_weights_and_biases(self, *losses: torch.Tensor) -> Tuple[int, List[float], List[float]]:
-        """
-        Implements Adaptive Multi-Losses Normalization with optimized PyTorch computations.
-
-        Args:
-            *losses (torch.Tensor): Variable number of loss tensors, one for each loss function.
-
-        Returns:
-            tuple: A tuple containing (number of losses, weights list, biases list).
-        """
-        # Ensure all losses are PyTorch tensors
-        losses = [loss if isinstance(loss, torch.Tensor) else torch.tensor(loss, dtype=torch.float64) 
-                 for loss in losses]
-        
-        num_losses = len(losses)
-        
-        if num_losses < 1:
-            raise ValueError("At least one loss array must be provided.")
-        
-        # Vectorized computation using PyTorch operations
-        means = torch.stack([torch.mean(loss) for loss in losses])
-        stds = torch.stack([torch.std(loss, unbiased=False) for loss in losses])
-        
-        # Handle zero standard deviations more efficiently
-        stds = torch.clamp(stds, min=self.eps)
-        if torch.any(stds == self.eps):
-            warnings.warn("A loss has zero standard deviation and was set to a small value.")
-        
-        ref_std = stds[0]
-        ref_mean = means[0]
-        
-        # Vectorized computation of weights and biases using PyTorch
-        weights = ref_std / stds
-        biases = ref_mean - weights * means
-        
-        return num_losses, weights.tolist(), biases.tolist()
-
-    def _compute_statistics(self, data: torch.Tensor) -> torch.Tensor:
-        """
-        Compute comprehensive statistics for a dataset efficiently using PyTorch.
-        
-        Args:
-            data (torch.Tensor): Input data tensor
-            
-        Returns:
-            torch.Tensor: Tensor of [mean, std, median, q25, q75]
-        """
-        return torch.tensor([
-            torch.mean(data),
-            torch.std(data),
-            torch.median(data),
-            torch.quantile(data, 0.25),
-            torch.quantile(data, 0.75)
-        ])
-
-    def plot_loss_distributions(self, losses: List[torch.Tensor], weights: List[float], 
-                               biases: List[float], loss_names: Optional[List[str]] = None, 
-                               save_path: str = 'plots', save_individual_plots: bool = True, 
-                               plot_dpi: int = 300, plot_format: str = 'png') -> None:
-        """
-        Create optimized box plots showing the distribution of losses before and after weighting.
-        Converts to NumPy only at the final plotting stage.
-        
-        Args:
-            losses (list of torch.Tensor): List of loss tensors for each loss function
-            weights (list): List of computed weights for each loss function
-            biases (list): List of computed biases for each loss function
-            loss_names (list, optional): Names of the loss functions for labeling
-            save_path (str): Directory to save the plots
-            save_individual_plots (bool): Whether to save individual loss function plots
-            plot_dpi (int): DPI for saved plots
-            plot_format (str): Format for saved plots ('png', 'pdf', 'svg', etc.)
-        """
-        # Create plots directory if it doesn't exist
-        os.makedirs(save_path, exist_ok=True)
-        
-        num_losses = len(losses)
-        if loss_names is None:
-            loss_names = [f'Loss_{i+1}' for i in range(num_losses)]
-        
-        # Pre-compute all weighted losses using PyTorch
-        before_weighting = losses
-        after_weighting = [weight * loss_array + bias 
-                          for loss_array, weight, bias in zip(losses, weights, biases)]
-        
-        # Compute all statistics using PyTorch
-        before_stats = torch.stack([self._compute_statistics(loss_array) for loss_array in before_weighting])
-        after_stats = torch.stack([self._compute_statistics(weighted_losses) for weighted_losses in after_weighting])
-        
-        # Convert to NumPy only for plotting
-        before_weighting_np = [loss.cpu().numpy() for loss in before_weighting]
-        after_weighting_np = [loss.cpu().numpy() for loss in after_weighting]
-        before_stats_np = before_stats.cpu().numpy()
-        after_stats_np = after_stats.cpu().numpy()
-        
-        # Create the main comparison plot
-        self._create_comparison_plot(before_weighting_np, after_weighting_np, before_stats_np, after_stats_np,
-                                   loss_names, weights, biases, save_path, plot_dpi, plot_format)
-        
-        # Create individual plots if requested
-        if save_individual_plots:
-            self._create_individual_plots(before_weighting_np, after_weighting_np, loss_names,
-                                        save_path, plot_dpi, plot_format)
-
-    def _create_comparison_plot(self, before_weighting: List[np.ndarray], 
-                               after_weighting: List[np.ndarray],
-                               before_stats: np.ndarray, after_stats: np.ndarray,
-                               loss_names: List[str], weights: List[float], 
-                               biases: List[float], save_path: str, 
-                               plot_dpi: int, plot_format: str) -> None:
-        """Create the main comparison plot efficiently."""
-        fig, (ax1, ax2) = plt.subplots(1, 2, figsize=(15, 6))
-        
-        # Plot 1: Before vs After box plots
-        all_data = []
-        all_labels = []
-        
-        for i, name in enumerate(loss_names):
-            all_data.extend([before_weighting[i], after_weighting[i]])
-            all_labels.extend([f'{name}\nBefore', f'{name}\nAfter'])
-        
-        bp1 = ax1.boxplot(all_data, labels=all_labels, patch_artist=True)
-        
-        # Color coding with predefined colors
-        colors = ['lightblue', 'darkblue', 'lightgreen', 'darkgreen', 'lightcoral', 'darkred', 
-                 'lightyellow', 'gold', 'lightpink', 'deeppink']
-        
-        for i, patch in enumerate(bp1['boxes']):
-            patch.set_facecolor(colors[i % len(colors)])
-            patch.set_alpha(0.7)
-        
-        ax1.set_title('Loss Distributions: Before vs After Weighting', fontsize=14, fontweight='bold')
-        ax1.set_ylabel('Loss Value', fontsize=12)
-        ax1.tick_params(axis='x', rotation=45)
-        ax1.grid(True, alpha=0.3)
-        
-        # Plot 2: Statistical comparison
-        num_losses = len(loss_names)
-        x = np.arange(num_losses)
-        width = 0.35
-        
-        # Plot means with error bars (std)
-        ax2.bar(x - width/2, before_stats[:, 0], width, label='Before Weighting', 
-                yerr=before_stats[:, 1], capsize=5, alpha=0.7, color='lightblue')
-        ax2.bar(x + width/2, after_stats[:, 0], width, label='After Weighting', 
-                yerr=after_stats[:, 1], capsize=5, alpha=0.7, color='darkblue')
-        
-        ax2.set_title('Statistical Comparison: Mean ± Std', fontsize=14, fontweight='bold')
-        ax2.set_xlabel('Loss Functions', fontsize=12)
-        ax2.set_ylabel('Loss Value', fontsize=12)
-        ax2.set_xticks(x)
-        ax2.set_xticklabels(loss_names, rotation=45)
-        ax2.legend()
-        ax2.grid(True, alpha=0.3)
-        
-        # Add weight and bias information as text
-        info_text = "Weights and Biases:\n"
-        for name, weight, bias in zip(loss_names, weights, biases):
-            info_text += f"{name}: w={weight:.4f}, b={bias:.4f}\n"
-        
-        plt.figtext(0.02, 0.02, info_text, fontsize=10, 
-                   bbox=dict(boxstyle="round,pad=0.3", facecolor="lightgray"))
-        
-        plt.tight_layout()
-        
-        # Save the plot
-        plot_filename = os.path.join(save_path, f'loss_distributions_comparison.{plot_format}')
-        plt.savefig(plot_filename, dpi=plot_dpi, bbox_inches='tight')
-        plt.close()
-        
-        logger.info(f"Loss distribution plots saved to: {plot_filename}")
-
-    def _create_individual_plots(self, before_weighting: List[np.ndarray], 
-                                after_weighting: List[np.ndarray],
-                                loss_names: List[str], save_path: str, 
-                                plot_dpi: int, plot_format: str) -> None:
-        """Create individual loss function plots efficiently."""
-        for loss_array, weighted_losses, name in zip(before_weighting, after_weighting, loss_names):
-            fig, ax = plt.subplots(figsize=(10, 6))
-            
-            # Create violin plot for better distribution visualization
-            violin_parts = ax.violinplot([loss_array, weighted_losses], positions=[1, 2])
-            
-            # Color the violin plots
-            violin_parts['bodies'][0].set_facecolor('lightblue')
-            violin_parts['bodies'][0].set_alpha(0.7)
-            violin_parts['bodies'][1].set_facecolor('darkblue')
-            violin_parts['bodies'][1].set_alpha(0.7)
-            
-            ax.set_title(f'{name}: Loss Distribution Before vs After Weighting', 
-                        fontsize=14, fontweight='bold')
-            ax.set_xlabel('Weighting', fontsize=12)
-            ax.set_ylabel('Loss Value', fontsize=12)
-            ax.set_xticks([1, 2])
-            ax.set_xticklabels(['Before', 'After'])
-            ax.grid(True, alpha=0.3)
-            
-            # Add statistics as text
-            before_mean, before_std = np.mean(loss_array), np.std(loss_array)
-            after_mean, after_std = np.mean(weighted_losses), np.std(weighted_losses)
-            
-            stats_text = f'Before: μ={before_mean:.4f}, σ={before_std:.4f}\nAfter: μ={after_mean:.4f}, σ={after_std:.4f}'
-            ax.text(0.02, 0.98, stats_text, transform=ax.transAxes, verticalalignment='top',
-                   bbox=dict(boxstyle="round,pad=0.3", facecolor="lightgray"))
-            
-            plt.tight_layout()
-            
-            # Save individual plot
-            safe_name = name.lower().replace(" ", "_").replace("/", "_")
-            individual_filename = os.path.join(save_path, f'{safe_name}_distribution.{plot_format}')
-            plt.savefig(individual_filename, dpi=plot_dpi, bbox_inches='tight')
-            plt.close()
-            
-            logger.info(f"Individual plot for {name} saved to: {individual_filename}")
-
-    def get_weighted_loss(self, individual_losses: List[torch.Tensor], 
-                         weights: List[float], biases: List[float]) -> torch.Tensor:
-        """
-        Compute the weighted combined loss for training.
-
-        Args:
-            individual_losses (List[torch.Tensor]): List of individual loss tensors
-            weights (List[float]): List of weights for each loss
-            biases (List[float]): List of biases for each loss
-            
-        Returns:
-            torch.Tensor: Combined weighted loss
-        """
-        if len(individual_losses) != len(weights) or len(individual_losses) != len(biases):
-            raise ValueError("Number of losses must match number of weights and biases")
-        
-        # Convert weights and biases to tensors for efficient computation
-        weights_tensor = torch.tensor(weights, device=individual_losses[0].device, 
-                                    dtype=individual_losses[0].dtype)
-        biases_tensor = torch.tensor(biases, device=individual_losses[0].device, 
-                                   dtype=individual_losses[0].dtype)
-        
-        # Stack losses and compute weighted sum efficiently
-        losses_tensor = torch.stack(individual_losses)
-        weighted_losses = weights_tensor * losses_tensor + biases_tensor
-        
-        return torch.sum(weighted_losses)
-
-if __name__ == '__main__':
-    from torch.utils.data import Dataset, DataLoader
-    import torch.nn as nn
-
-    # 1. Create a dummy dataset and DataLoader to simulate the real use case
-    class CustomDummyDataset(Dataset):
-        def __init__(self, num_samples=100, batch_size=8):
-            self.num_samples = num_samples
-            self.batch_size = batch_size
-        
-        def __len__(self):
-            return self.num_samples
-
-        def __getitem__(self, idx):
-            # ((motion_before, motion, motion_after), free)
-            return (torch.randn(1, 64, 64), torch.randn(1, 64, 64), torch.randn(1, 64, 64)), torch.randn(1, 64, 64)
-
-    batch_size = 8
-    dummy_dataset = CustomDummyDataset(num_samples=100, batch_size=batch_size)
-    dummy_dataloader = DataLoader(dummy_dataset, batch_size=batch_size)
-
-    # 2. Define loss functions
-    l1_loss = nn.L1Loss()
-    mse_loss = nn.MSELoss()
-    losses = [l1_loss, mse_loss]
-
-    # 3. Use the optimized adaptive loss normalization class
-    adaptive_loss_calculator = AdaMultiLossesNormTorch(eps=1e-8)
-    
-    print("=== PyTorch-Optimized Adaptive Multi-Loss Normalization ===")
-    print(f"Using epsilon: {adaptive_loss_calculator.eps}")
-
-    # Compute initial losses over the dataset with optimized PyTorch processing
-    initial_losses = adaptive_loss_calculator.compute_losses(
-        dataset=dummy_dataloader,
-        batch_size=batch_size,
-        losses=losses,
-        device='cpu'  # Use 'cuda' if available
-    )
-    
-    # Now working with PyTorch tensors instead of NumPy arrays
-    l1_losses_over_dataset = initial_losses[0]  # First row: L1 losses
-    mse_losses_over_dataset = initial_losses[1]  # Second row: MSE losses
-    
-    print(f"Computed L1 losses for {l1_losses_over_dataset.shape[0]} batches.")
-    print(f"Computed MSE losses for {mse_losses_over_dataset.shape[0]} batches.")
-    print(f"Loss tensor shape: {initial_losses.shape}")
-    print(f"Data type: {initial_losses.dtype}")
-
-    # Compute weights and biases from these initial losses with optimized PyTorch computation
-    num_losses, weights, biases = adaptive_loss_calculator.compute_normalized_weights_and_biases(
-        l1_losses_over_dataset, mse_losses_over_dataset
-    )
-    
-    print(f"\nNumber of losses: {num_losses}")
-    print(f"Computed weights: {weights}")
-    print(f"Computed biases: {biases}")
-    
-    # 4. Example of how to use the optimized weighted loss computation
-    print("\n--- PyTorch-Optimized Training Loop Example ---")
-    
-    motion_data_batch, free_batch = next(iter(dummy_dataloader))
-    
-    # Compute individual losses
-    batch_l1 = l1_loss(motion_data_batch[1], free_batch)
-    batch_mse = mse_loss(motion_data_batch[1], free_batch)
-    
-    # Use the optimized weighted loss computation
-    individual_losses = [batch_l1, batch_mse]
-    total_loss = adaptive_loss_calculator.get_weighted_loss(individual_losses, weights, biases)
-    
-    print(f"Batch L1 Loss: {batch_l1.item():.4f}")
-    print(f"Batch MSE Loss: {batch_mse.item():.4f}")
-    print(f"Weighted L1 part: {weights[0] * batch_l1.item() + biases[0]:.4f}")
-    print(f"Weighted MSE part: {weights[1] * batch_mse.item() + biases[1]:.4f}")
-    print(f"Total Combined Loss (PyTorch optimized): {total_loss.item():.4f}")
-    
-    # Verify the result matches manual computation
-    manual_total = (weights[0] * batch_l1 + biases[0]) + (weights[1] * batch_mse + biases[1])
-    print(f"Manual computation verification: {manual_total.item():.4f}")
-    print(f"Difference: {abs(total_loss.item() - manual_total.item()):.2e}")
-
-    # 5. Create optimized loss distribution plots (converts to NumPy only at the end)
-    print("\n--- Creating PyTorch-Optimized Loss Distribution Plots ---")
-    adaptive_loss_calculator.plot_loss_distributions(
-        losses=[l1_losses_over_dataset, mse_losses_over_dataset],
-        weights=weights,
-        biases=biases,
-        loss_names=['L1 Loss', 'MSE Loss'],
-        save_path='plots',
-        save_individual_plots=True,
-        plot_dpi=300,
-        plot_format='png'
-    )
-    
-    print("\n=== PyTorch Optimization Summary ===")
-    print("✓ Eliminated unnecessary NumPy conversions")
-    print("✓ All computations now use PyTorch tensors")
-    print("✓ Reduced memory transfers between CPU/GPU")
-    print("✓ Faster tensor operations with PyTorch")
-    print("✓ NumPy conversion only at final plotting stage")
-    print("✓ Better GPU utilization for large datasets")
-    print("✓ Maintained backward compatibility")
-    print("✓ Improved type safety with tensor operations")
